+++ CHANGELOG.rst
@@ -12,9 +12,16 @@ Unreleased
 
 Added
 -----
+- The unit cells and atomic positions of ``Phase`` objects can now be visualized
+  using ``Phase.plot_unit_cell``.
+- ``Phase`` objects can now be created using the seven crystal class methods ``Phase.triclinic``,
+  ``Phase.monoclinic``, ``Phase.orthorhombic``, ``Phase.tetragonal``, ``Phase.trigonal``, 
+  ``Phase.hexagonal``, and ``Phase.cubic``
 
 Changed
 -------
+- The IPF color keys can take 'x', 'y', and 'z' text strings as direction inputs
+  instead of requiring Vector3d objects.
 
 Removed
 -------